User update and language table creation
